refactor: split daemon_cmds.py into daemon/ package
Split the 486-line daemon_cmds.py into a well-organized package:

**Package structure:**
- `daemon/__init__.py` - Public API exports
- `daemon/_helpers.py` - Internal: _clear_session_files
- `daemon/lifecycle_cmds.py` - Daemon lifecycle (start, stop, pause, resume)
- `daemon/health_cmds.py` - Health monitoring (diagnose)
- `daemon/device_mgmt_cmds.py` - Device registry (devices, add, remove)

**Test updates:**
- Updated all test imports from `eab.cli.daemon_cmds` → `eab.cli.daemon`
- Fixed subprocess/function patches to point to lifecycle_cmds module
- Updated _clear_session_files imports to daemon._helpers

**Results:**
- ✅ All 60 daemon tests passing
- ✅ Clean separation by concern (lifecycle/health/devices)
- ✅ Public API preserved in __init__.py
- ✅ No breaking changes
